Replace "Copy Headers" phase with "Copy Files" for libochamcrest target.
Static library targets ignore the Copy Headers phase, so it wasn't doing anything. But to be able to reference the headers by a logical path ("<OCHamhcrest/xxx.h>"), we need to to copy them to the appropriate location in the Build directory, because it's rather a challenge to set up an equivalent path to the original source files when the project is included as a submodule of another project's repository.